Logic
This is the logic part of the game, where the game is actually running and checking the player inputs. It checks what would happen if the game is to end and there is a winner or a draw. What are the possible spots for the user or ai to pick to win and to check if the spot is taken and the creation of the tictactoe board and assigning the character to the spot the user has chosen, all happens with this code.